+++ app/src/main/java/com/agileburo/anytype/ui/page/PageFragment.kt
@@ -622,11 +622,17 @@ open class PageFragment :
                 }
                 is Command.OpenTurnIntoPanel -> {
                     TurnIntoFragment.single(
-                        target = command.target
+                        target = command.target,
+                        excludedCategories = command.excludedCategories,
+                        excludedTypes = command.excludedTypes
+
                     ).show(childFragmentManager, null)
                 }
                 is Command.OpenMultiSelectTurnIntoPanel -> {
-                    TurnIntoFragment.multiple().show(childFragmentManager, null)
+                    TurnIntoFragment.multiple(
+                        excludedTypes = command.excludedTypes,
+                        excludedCategories = command.excludedCategories
+                    ).show(childFragmentManager, null)
                 }
                 is Command.OpenBookmarkSetter -> {
                     CreateBookmarkFragment.newInstance(